socket-server-pc/requirements.txt
python-uinput==0.11.2
websocket-client==0.53.0
